requirements.txt
flask==2.3.3
flask-socketio==5.3.6
google-api-python-client==2.86.0
python-dotenv==1.0.0
mysql-connector-python==8.0.33
python-dateutil==2.8.2
eventlet==0.33.3
gunicorn==20.1.0
werkzeug==2.3.7 